Add files via upload
This project is an RNG using the cryptographic algorithm yarrow. It can generate a random number from 16 bytes to 2048 bytes. The Yarrow algorithm is used to generate random numbers to use in encryption and hash functions, as well as this code, will check whether the generated number we are generating is unique or not. This project is done in Python language.